requirements.txt
python-dotenv
openai
bertopic
scikit-learn
nltk
streamlit
